+++ database.py
@@ -1,10 +1,11 @@
+import os
 from datetime import datetime
 from sqlalchemy import Column, Integer, String, Text, DateTime
 from sqlalchemy.ext.asyncio import create_async_engine, async_sessionmaker
 from sqlalchemy.orm import DeclarativeBase
 from dotenv import load_dotenv
+
 load_dotenv()
-import os
 
 host = os.getenv('host')
 port = os.getenv('port')
@@ -12,15 +13,17 @@ user = os.getenv('user')
 password = os.getenv('password')
 database = os.getenv('database')
 
-DATABASE_URL = f"postgresql+asyncpg://{user}:{password}@{host}/{database}"
+DATABASE_URL = f"postgresql+asyncpg://{user}:{password}@{host}/{database}"  # маршрут в основной проект
 
 engine = create_async_engine(url=DATABASE_URL)
 new_session = async_sessionmaker(engine, expire_on_commit=False)
 
+
 class Model(DeclarativeBase):
     id = Column(Integer, primary_key=True, autoincrement=True)
 
-class DAccident(Model):
+
+class DAccident(Model):                                 # большую часть классов можно взять с service
     __tablename__ = "accident_accident"
     number = Column(String(10))
     category = Column(String(20))
@@ -37,6 +40,7 @@ class DAccident(Model):
     decide = Column(String(2000))
     status = Column(String(5))
 
+
 class DGazprom(Model):
     __tablename__ = "gazprom_gazprom"
     ip = Column(String(15))
@@ -47,10 +51,12 @@ class DGazprom(Model):
     comment = Column(String(500))
     geo = Column(String(30))
 
+
 class DManual(Model):
     __tablename__ = "manual_manual"
     model = Column(String(255))
     description = Column(Text)
+
 
 class DUser(Model):
     __tablename__ = "_userbot"
@@ -62,17 +68,20 @@ class DUser(Model):
     email = Column(String(50))
     tg_id = Column(String(50))
 
+
 class DVisitedUser(Model):
     __tablename__ = "_visited_users"
     login = Column(String(30))
     date_created = Column(DateTime(timezone=False))
     action = Column(String(50))
 
+
 class DBaseStation(Model):
     __tablename__ = "base_station_basestation"
     comment = Column(String(255))
     address = Column(String(255))
     number = Column(Integer())
+
 
 class DAllInfo(Model):
     __tablename__ = "fttx_fttx"
@@ -82,6 +91,7 @@ class DAllInfo(Model):
     number = Column(String(10))
     description = Column(Text(1000))
     askue = Column(Integer)
+
 
 class DAddInfo(Model):
     __tablename__ = "info_info"
