+++ repository.py
@@ -1,26 +1,22 @@
 from datetime import datetime
-from sqlalchemy import select, insert, update
+from sqlalchemy import select, insert, update, and_
 from database import (DGazprom, DManual, DUser, DVisitedUser, DBaseStation,
                       DAllInfo, DAccident, DAddInfo, new_session)
 
 class Repo:
     @classmethod
     async def select_pass(cls, login, psw, tg_id):
-        print(psw,"\n")
+        """Authentication"""
         async with new_session() as session:
-            tg_id = str(tg_id)
-            pswrd = psw.decode("UTF-8")
-            password = str(pswrd)
-            q = select(DUser).where(DUser.login == login, DUser.password == password,
-                                    DUser.tg_id == tg_id)
+            q = select(DUser).where(DUser.login == login, DUser.password == psw,
+                                    DUser.tg_id == str(tg_id))
             result = await session.execute(q)
             answer = result.scalar()
-            await session.close()
             return answer
-
 
     @classmethod
     async def select_accident(cls, status):
+        """Select accident by status"""
         async with new_session() as session:
             if status == "open" or status == "check":
                 q = select(DAccident).order_by(DAccident.id.desc()).where(DAccident.status == status)
@@ -29,83 +25,80 @@ class Repo:
             result = await session.execute(q)
             answer = result.scalars()
             await session.commit()
-            await session.close()
             return answer
-
 
     @classmethod
     async def select_accident_number(cls, number):
+        """Select accident by number"""
         async with new_session() as session:
             number = str(number)
             query = select(DAccident).where(DAccident.number == number)
             result = await session.execute(query)
             answer = result.scalar()
             await session.commit()
-            await session.close()
             return answer
-
 
     @classmethod
     async def select_azs(cls, number):
+        """Select AZS"""
         async with new_session() as session:
             number = str(number)
             q = select(DGazprom).where(DGazprom.number == number)
             result = await session.execute(q)
             answer = result.scalar()
-            await session.close()
             return answer
 
     @classmethod
     async def select_manual(cls, ssid):
+        """Select manual"""
         async with new_session() as session:
             q = select(DManual).where(DManual.id == int(ssid))
             result = await session.execute(q)
             answer = result.scalar()
-            await session.close()
             return answer
 
-    # # insert into _visited_users
     @classmethod
     async def insert_into_visited_date(cls, login, action):
+        """Insert visited users"""
         async with new_session() as session:
             date_created = datetime.now()
             q = DVisitedUser(login=login, date_created=date_created, action=action)
             session.add(q)
             await session.commit()
-            await session.close()
             return
-    #
+
     @classmethod
     async def select_action(cls, number):
+        """View visited users"""
         async with new_session() as session:
             query = select(DVisitedUser).order_by(DVisitedUser.id.desc()).limit(int(number))
             result = await session.execute(query)
             answer = result.scalars().all()
             return answer
-    #
-    #
+
     @classmethod
     async def select_bs_number(cls, number):
+        """Select BS by number"""
         async with new_session() as session:
             query = select(DBaseStation).where(DBaseStation.number == int(number))
             result = await session.execute(query)
             answer = result.scalar()
             await session.commit()
-            await session.close()
             return answer
 
     @classmethod
     async def select_bs_address(cls, address):
+        """Select BS by street"""
         async with new_session() as session:
             query = select(DBaseStation).where(DBaseStation.address.like(f"%{address}%"))
             result = await session.execute(query)
             answer = result.scalars()
             await session.commit()
-            await session.close()
             return answer
 
     @classmethod
     async def select_all_info(cls, temp):
+        """Select info fttx"""
         async with new_session() as session:
             result = temp.split(", ")
             city = str(result[0])
@@ -117,20 +110,20 @@ class Repo:
             result = await session.execute(query)
             answer = result.scalar()
             await session.commit()
-            await session.close()
             return answer
 
     @classmethod
     async def select_stat(cls):
+        """Select stat visited"""
         async with new_session() as session:
             q = select(DVisitedUser).order_by(DVisitedUser.id.desc()).limit(5)
             result = await session.execute(q)
             answer = result.scalars().all()
-            await session.close()
             return answer
 
     @classmethod
     async def insert_info(cls, info):
+        """insert new info in fttx_fttx"""
         async with new_session() as session:
             reestr = int(info[0])
             date = datetime.now()
@@ -154,6 +147,7 @@ class Repo:
 
     @classmethod
     async def update_accident(cls, l):
+        """Update accident by number"""
         async with new_session() as session:
             number = str(l[0])
             decide = str(l[2])
@@ -163,16 +157,29 @@ class Repo:
             )
             answer = await session.execute(q)
             await session.commit()
-            await session.close()
             return answer
 
     @classmethod
     async def grafik_ring(cls, start_date, end_date):
+        """Select date for graph"""
         async with new_session() as session:
             query = select(DAddInfo).where(
                 DAddInfo.date_created.between(start_date, end_date)
             )
             result = await session.execute(query)
             answer = result.scalars().all()
+            await session.commit()
+            return answer
+
+    @classmethod
+    async def exit_user_bot(cls, tg_id):
+        """Exit."""
+        async with new_session() as session:
+            q = select(DUser).where(and_(DUser.tg_id == str(tg_id)))
+            result = await session.execute(q)
+            answer = result.scalars().first()
+            if answer is None:
+                return None
+            answer.active = 0
             await session.commit()
             return answer
